requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.15.2
gunicorn==21.2.0
pillow==12.0.0
python-dotenv==1.2.1
qrcode==8.2
sqlparse==0.5.3
whitenoise==6.11.0
django-import-export==4.3.0
openpyxl==3.1.3
django-filter==23.2